fix(snapshot): finish int16_t->int32_t widening on three more call sites
The earlier yy_shift_ofst/yy_reduce_ofst widening missed three
helpers that also dereference the snapshot tables.  All three were
working on small grammars but would silently mis-index on PG-class
inputs:

  - src/parse_context.c::snap_find_shift_action
  - src/parse_context.c::snap_find_reduce_action
    (used by the disambiguation / extension pipeline -- not the
    main runtime engine, which is why the previous round did not
    flush them out)

  - src/jit_context.c::table_find_shift_action
    (used by jit_parse_batch's interpreted fallback when the
    monolithic JIT function is absent -- now reachable, since the
    large-grammar codegen path skips emitting jit_parse_sequence)

  - src/glr.c::find_shift_action / find_reduce_action and the
    public glr_parser_feed signature in include/glr.h
    (no internal callers, but the public-API typing was wrong --
    a future caller passing snap->yy_shift_ofst would have hit
    an int32_t* vs int16_t* mismatch)

These were caught while diagnosing why state-after-SELECT in the
PG snapshot rejects every common token (the test corpus only
accepts INSERT statements).  That investigation showed the
snapshot is internally consistent at the bit level: state 0 +
SELECT shifts to state 1573, but state 1573 has only one explicit
non-default action (lookahead=38), with yy_default[1573] =
yy_error_action.  That looks like a real grammar artefact in the
converted PG gram.lime -- not a runtime engine bug -- and is
tracked separately (see the comment in tests/test_pg_grammar.c).

75 lime tests pass.

+++ include/glr.h
@@ -110,7 +110,7 @@ void glr_parser_destroy(GLRParser *parser);
 */
 int glr_parser_feed(GLRParser *parser, uint16_t token, const uint16_t *yy_action,
                     uint32_t action_count, const uint16_t *yy_lookahead, uint32_t lookahead_count,
-                    const int16_t *yy_shift_ofst, const int16_t *yy_reduce_ofst,
+                    const int32_t *yy_shift_ofst, const int32_t *yy_reduce_ofst,
                     const uint16_t *yy_default, uint32_t nstate, const uint16_t *yy_rule_lhs,
                     const int8_t *yy_rule_nrhs, uint32_t nrule, uint16_t min_shiftreduce,
                     uint16_t max_shiftreduce, uint16_t min_reduce, uint16_t max_reduce,
